Fixed tons of 32-bit warnings and also added Get() -> size_t limit tests to all Get's of references

+++ test/testarray.cpp
@@ -718,7 +718,7 @@ TEST(FindSSE) {
 		a.Add(10000);
 	}
 
-	for(uint64_t i = 0; i < 100; i++) {
+	for(size_t i = 0; i < 100; i++) {
 		a.Set(i, 123);
 		size_t t = a.Find(123);
 		assert(t == i);
@@ -823,7 +823,7 @@ TEST(Greater) {
 	{
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(0);
 		}
 		size_t t = a.Query<GREATER>(0, 0, (size_t)-1);
@@ -831,10 +831,10 @@ TEST(Greater) {
 
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(0);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 1);
 			size_t t = a.Query<GREATER>(0, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -842,10 +842,10 @@ TEST(Greater) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(2);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 3);
 			size_t t = a.Query<GREATER>(2, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -853,10 +853,10 @@ TEST(Greater) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(10);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 11);
 			size_t t = a.Query<GREATER>(10, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -864,20 +864,20 @@ TEST(Greater) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(100);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 110);
 			size_t t = a.Query<GREATER>(100, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 100);
 		}
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(200);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 210);
 			size_t t = a.Query<GREATER>(200, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -885,21 +885,21 @@ TEST(Greater) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(10000);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 11000);
 			size_t t = a.Query<GREATER>(10000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 10000);
 		}
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(40000);
 		}
 
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 41000);
 			size_t t = a.Query<GREATER>(40000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -907,10 +907,10 @@ TEST(Greater) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(1000000);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 1100000);
 			size_t t = a.Query<GREATER>(1000000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -918,10 +918,10 @@ TEST(Greater) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(1000ULL*1000ULL*1000ULL*1000ULL);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 1000ULL*1000ULL*1000ULL*1000ULL + 1ULL);
 			size_t t = a.Query<GREATER>(1000ULL*1000ULL*1000ULL*1000ULL, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -944,7 +944,7 @@ TEST(Less) {
 	{
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(0);
 		}
 		size_t t = a.Query<LESS>(0, 0, (size_t)-1);
@@ -952,10 +952,10 @@ TEST(Less) {
 
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(1);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 0);
 			size_t t = a.Query<LESS>(1, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -963,10 +963,10 @@ TEST(Less) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(3);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 2);
 			size_t t = a.Query<LESS>(3, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -974,10 +974,10 @@ TEST(Less) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(11);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 10);
 			size_t t = a.Query<LESS>(11, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -985,20 +985,20 @@ TEST(Less) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(110);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 100);
 			size_t t = a.Query<LESS>(110, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 110);
 		}
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(210);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 200);
 			size_t t = a.Query<LESS>(210, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -1006,21 +1006,21 @@ TEST(Less) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(11000);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 10000);
 			size_t t = a.Query<LESS>(11000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
 			a.Set(i, 11000);
 		}
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(41000);
 		}
 
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 40000);
 			size_t t = a.Query<LESS>(41000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -1028,10 +1028,10 @@ TEST(Less) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(1100000);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 1000000);
 			size_t t = a.Query<LESS>(1100000, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
@@ -1039,10 +1039,10 @@ TEST(Less) {
 		}
 
 		a.Clear();	
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Add(1000ULL*1000ULL*1000ULL*1000ULL);
 		}
-		for(int i = 0; i < items; i++) {
+		for(size_t i = 0; i < items; i++) {
 			a.Set(i, 1000ULL*1000ULL*1000ULL*1000ULL - 1ULL);
 			size_t t = a.Query<LESS>(1000ULL*1000ULL*1000ULL*1000ULL, 0, (size_t)-1);
 			CHECK_EQUAL(i, t);
